+++ test/core-model-test/src/main/java/com/ainrif/apiator/test/model/core/Dto00_MethodReturnDto.java
@@ -1,0 +1,8 @@
+package com.ainrif.apiator.test.model.core;
+
+/**
+ * This Dto must be present only as method return value to check that kind of introspection
+ */
+public class Dto00_MethodReturnDto {
+    public String stringField;
+}